Rename geoclimateChain to workflow
